requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
flake8==3.8.4
Flask==1.1.2
Flask-API==2.0
idna==2.10
importlib-metadata==3.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
py4j==0.10.9.1
pycodestyle==2.6.0
pyflakes==2.2.0
pypmml==0.9.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
rope==0.18.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn-pandas==2.0.3
sklearn2pmml==0.64.0
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
